Refactoring linkedlist.cpp
Motivation:
Currently the methods in this class are named more like a stack which
was not the original intention.
The intention is still to learn the basic of implementing a singly
linked list I'm not using templates just to keep things simple.

Modifications:
Renamed a push and pop methods and added an empty method.

Result:
All test pass:
make check