Initializing the parentVertex array to -1.
Motivation:
Currently only the entry zero was initilized to -1 which made it look
like the all entries that did have explicit parents sets, looked like
their parent was zero. This was due to a mistake on my part where I
thought that it was possible to initialize a complete array using the
{-1} syntax. This only sets the first element to -1 and the rest to 0.

Modifications:
Added a look to initilizes all entries to -1. This would probably be
possible to use something like std::fill_n(array, 20, -1) from
<algorithm> but I've not gotten that far yet.

Result:
The complete parentVertex array is initialized to -1.